requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.6.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
stripe>=8.7.0
sendgrid>=6.11.0
reportlab>=4.1.0
python-multipart>=0.0.9
boto3>=1.34.0
apscheduler>=3.10.0
python-dateutil>=2.9.0
pytest>=8.0.0
httpx>=0.27.0
